+++ crates/forge/tests/e2e/mod.rs
@@ -6,6 +6,7 @@ mod build_trace_data;
 mod clean;
 mod collection;
 mod color;
+mod completions;
 mod components;
 mod contract_artifacts;
 #[cfg(not(target_os = "windows"))]